+++ httpcore/_sync/connection_pool.py
@@ -5,7 +5,6 @@ from typing import Iterable, Iterator, List, Optional, Type
 
 from .._exceptions import ConnectionNotAvailable, UnsupportedProtocol
 from .._models import Origin, Request, Response
-from .._ssl import default_ssl_context
 from .._synchronization import Event, Lock
 from ..backends.sync import SyncBackend
 from ..backends.base import NetworkBackend
@@ -82,9 +81,6 @@ class ConnectionPool(RequestInterface):
             uds: Path to a Unix Domain Socket to use instead of TCP sockets.
             network_backend: A backend instance to use for handling network I/O.
         """
-        if ssl_context is None:
-            ssl_context = default_ssl_context()
-
         self._ssl_context = ssl_context
 
         self._max_connections = (